% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{get_cmdstan_model}
\alias{get_cmdstan_model}
\title{Get a compiled cmdstanr model from inst/stan}
\usage{
get_cmdstan_model(stan_file)
}
\description{
Get a compiled cmdstanr model from inst/stan
}
\keyword{internal}
