% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-utils.R
\name{plateau_gaussian_density_left}
\alias{plateau_gaussian_density_left}
\title{Computes the normalised plateau–Gaussian density used for y = 0}
\usage{
plateau_gaussian_density_left(x, mu, sigma, L)
}
\arguments{
\item{x}{vector of x-values}

\item{mu}{mean location of the Gaussian tail}

\item{sigma}{transition spread}

\item{L}{lower truncation}
}
\description{
Computes the normalised plateau–Gaussian density used for y = 0
}
