% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-utils.R
\name{plateau_gaussian_density_right}
\alias{plateau_gaussian_density_right}
\title{Computes the normalised plateau–Gaussian density used for y = 1}
\usage{
plateau_gaussian_density_right(x, mu, sigma, U)
}
\arguments{
\item{x}{vector of x-values}

\item{mu}{mean of the Gaussian tail}

\item{sigma}{transition spread}

\item{U}{upper truncation}
}
\description{
Computes the normalised plateau–Gaussian density used for y = 1
}
