% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_priors.R
\name{stage_priors}
\alias{stage_priors}
\title{Construct prior hyperparameters for the STAGE model}
\usage{
stage_priors(
  x = NULL,
  y = NULL,
  prior_mu_m50_mu = if (!is.null(x)) stats::median(x) else 0,
  prior_mu_m50_tau = if (!is.null(x)) diff(range(x))/6 else 100,
  prior_d_mu = if (!is.null(x)) diff(range(x))/4 else 0,
  prior_d_tau = if (!is.null(x)) diff(range(x))/4 else 100,
  prior_sigma_x_tau = if (!is.null(x)) stats::sd(x) else 100,
  prior_sigma_alpha_tau = if (!is.null(x)) diff(range(x))/6 else 100
)
}
\arguments{
\item{x, y}{Optional data vectors to set sensible defaults.}

\item{prior_mu_m50_mu}{Mean of the prior on the global m50.}

\item{prior_mu_m50_tau}{SD of the prior on the global m50.}

\item{prior_d_mu}{Mean of the prior on d.}

\item{prior_d_tau}{SD of the prior on d.}

\item{prior_sigma_x_tau}{SD scale for sigma_x.}

\item{prior_sigma_alpha_tau}{SD scale for sigma_alpha (between-pop).}
}
\value{
A list with elements matching the Stan data block.
}
\description{
Construct prior hyperparameters for the STAGE model
}
