% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_point.R
\name{transition_point}
\alias{transition_point}
\title{Generic transition point extractor}
\usage{
transition_point(object, ...)
}
\arguments{
\item{object}{A model fit object.}

\item{...}{Further arguments passed to methods.}
}
\value{
Typically a numeric vector or list summarising the transition point.
}
\description{
Generic transition point extractor
}
