% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_stan_utils.R, R/stan_utils.R
\name{get_cmdstan_model}
\alias{get_cmdstan_model}
\title{Get compiled cmdstanr model}
\usage{
get_cmdstan_model(stan_file)

get_cmdstan_model(stan_file)
}
\description{
Get compiled cmdstanr model

Get a compiled cmdstanr model from inst/stan
}
\keyword{internal}
